Add files via upload
These are the files for the project called BlueJacket. Blue Jacket is a big company that sells branded jackets with the best quality in the world. All jackets in Blue Jacket are purchased from the best vendor around the world that has a great reputation in this industry. Blue Jacket also sell these jackets to their customer around the globe and always make their customer happy. I am assigned to create the database and enter the data query into the database.